+++ lib/hallmonitor/monitored.rb
@@ -9,16 +9,18 @@ module Hallmonitor
       # @options [Hash] Optional settings:
       #   metric_name: [String] Metric name to emit, defaults to
       #   "#{underscore(name)}.#{method_sym}"
-      def timer_for(method_sym, options = {})
-        metric_name = options[:metric_name] || "#{underscore(name)}.#{method_sym}"
-        send(:define_method, "#{method_sym}_with_hallmonitor_timer") do |*args|
-          watch(metric_name) do
-            send("#{method_sym}_without_hallmonitor_timer".to_sym, *args)
+      def timer_for(method_sym, metric_name: nil, tags: {})
+        metric_name ||= "#{underscore(name)}.#{method_sym}"
+        undecorated_method_sym = "#{method_sym}_without_hallmonitor_timer".to_sym
+        decorated_method_sym = "#{method_sym}_with_hallmonitor_timer".to_sym
+        send(:define_method, decorated_method_sym) do |*args|
+          watch(metric_name, tags: tags) do
+            send(undecorated_method_sym, *args)
           end
         end
 
-        alias_method "#{method_sym}_without_hallmonitor_timer".to_sym, method_sym
-        alias_method method_sym, "#{method_sym}_with_hallmonitor_timer".to_sym
+        alias_method undecorated_method_sym, method_sym
+        alias_method method_sym, decorated_method_sym
       end
 
       # Sets up a counter for a method by symbol.  Method must have already been
@@ -27,17 +29,20 @@ module Hallmonitor
       # @options [Hash] Optional settings:
       #   metric_name: [String] Metric name to emit, defaults to
       #   "#{underscore(name)}.#{method_sym}"
-      def count_for(method_sym, options = {})
-        metric_name = options[:metric_name] || "#{underscore(name)}.#{method_sym}"
-        send(:define_method, "#{method_sym}_with_hallmonitor_counter") do |*args|
-          emit(metric_name)
-          send("#{method_sym}_without_hallmonitor_counter".to_sym, *args)
+      def count_for(method_sym, metric_name: nil, tags: {})
+        metric_name ||= "#{underscore(name)}.#{method_sym}"
+        undecorated_method_sym = "#{method_sym}_without_hallmonitor_counter".to_sym
+        decorated_method_sym = "#{method_sym}_with_hallmonitor_counter".to_sym
+        send(:define_method, decorated_method_sym) do |*args|
+          emit(metric_name, tags: tags)
+          send(undecorated_method_sym, *args)
         end
 
-        alias_method "#{method_sym}_without_hallmonitor_counter".to_sym, method_sym
-        alias_method method_sym, "#{method_sym}_with_hallmonitor_counter".to_sym
+        alias_method undecorated_method_sym, method_sym
+        alias_method method_sym, decorated_method_sym
       end
 
+      # :reek:UtilityFunction
       def underscore(value)
         word = value.dup
         word.gsub!(/::/, '.')
@@ -51,17 +56,19 @@ module Hallmonitor
 
     def self.included(base)
       base.extend(ClassMethods)
+      base.extend(self)
     end
 
-    # Emits an event: self if the event param is nil, the passed in event if
-    # it's an {Event}, or constructs a {Event} from the passed in param.
+    # Emits an event: self if the event param is nil, the passed in
+    # event if it's an {Event}, or constructs a {Event} from the
+    # passed in param.
     #
-    # If the parameter is a {Hallmonitor::Event}, it will be emitted as is.
-    # Otherwise, a new {Hallmonitor::Event} will be created with the parameter
-    # and emitted.
+    # If the parameter is a {Hallmonitor::Event}, it will be emitted
+    # as is.  Otherwise, a new {Hallmonitor::Event} will be created
+    # with the parameter and emitted.
+    #
     # @param event [Mixed] The thing to emit, see method description
     # @yield [to_emit] The thing that's going to be emitted
-    # @return nil
     def emit(event = nil, tags: {})
       to_emit = self
       unless event.nil?